+++ lib/ductwork/models/result.rb
@@ -9,6 +9,7 @@ module Ductwork
     enum :result_type,
          success: "success",
          failure: "failure",
-         timed_out: "timed_out"
+         timed_out: "timed_out",
+         process_crashed: "process_crashed"
   end
 end